% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSNE.R
\docType{class}
\name{DataScale-class}
\alias{DataScale-class}
\alias{DataScale}
\title{A Reference Class to represent a DataScale.}
\description{
A Reference Class to represent a DataScale.
}
\section{Fields}{

\describe{
\item{\code{tmatix}}{A matrix of shape num_scalesXnum_data_points.}

\item{\code{size}}{A numeric vale of points in scale.}

\item{\code{datapoints}}{A vector of all the datapoints represented by numeric values.}

\item{\code{scalenum}}{A numeric value of the current scale number in HSNE object.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(num_scales, tmatrix)}}{Initialize data-scale.}

\item{\code{toString()}}{This method returns the string representation of a data-scale.}
}}

