% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSNE.R
\docType{class}
\name{HSNE-class}
\alias{HSNE-class}
\alias{HSNE}
\title{A Reference Class to represent a HSNE.}
\description{
A Reference Class to represent a HSNE.
}
\section{Fields}{

\describe{
\item{\code{num_scales}}{A numeric value of the number of scales.}

\item{\code{scales}}{A data frame with DataScale and SubScale.}

\item{\code{index}}{A numeric value of the current scale number.}
}}

\section{Methods}{

\describe{
\item{\code{cluster_scale(scalenumber, prop_method = "cluster")}}{Cluster the given scale using Leiden community detection.}

\item{\code{getItem(index)}}{Get item from scale.}

\item{\code{get_datascale_mappings(scalenumber)}}{Get datascale of choice at scalenumber. Used automatically by cluster_scale(scalenumber) method.}

\item{\code{get_map_by_cluster(scalenumber, clustering)}}{Get map for prop_method=clustering for a subscale at scalenumber. Used automatically by cluster_scale(scalenumber) method.}

\item{\code{get_topscale()}}{Get the top scale (data scale).}

\item{\code{initialize(num_scales)}}{Initialize a HSNE with num_scales scales.}

\item{\code{iter()}}{Iterate scale.}

\item{\code{next_scale()}}{Get next scale.}

\item{\code{run_leiden(scalenumber)}}{Run leiden algorithm on a scale}

\item{\code{setItem(index, value)}}{Set item on scale.}

\item{\code{toString()}}{To string for HSNE.}
}}

