% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSNE.R
\docType{class}
\name{Logger-class}
\alias{Logger-class}
\alias{Logger}
\title{A Reference Class to represent a Logger}
\description{
Printin messages to console
}
\section{Fields}{

\describe{
\item{\code{enabled}}{A boolean to enable printing.}
}}

\section{Methods}{

\describe{
\item{\code{disable()}}{Disable Logger.}

\item{\code{enable()}}{Enable Logger.}

\item{\code{initialize(enabled)}}{Initialize a Logger.}

\item{\code{log(message)}}{Show message.}
}}

