% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSNE.R
\docType{class}
\name{SubScale-class}
\alias{SubScale-class}
\alias{SubScale}
\title{A Reference Class to represent a SubScale}
\description{
A Reference Class to represent a SubScale
}
\section{Fields}{

\describe{
\item{\code{tmatix}}{A matrix of shape num_scalesXnum_data_points.}

\item{\code{size}}{A numeric vale of points in scale.}

\item{\code{datapoints}}{A vector of all the datapoints represented by numeric values.}

\item{\code{scalenum}}{A numeric value of the current scale number in HSNE object.}

\item{\code{num_scales}}{A numeric value of the total number scales in HSNE object.}

\item{\code{lm_to_original}}{A vector of labels from each datapoints to its closest original landmarks.}

\item{\code{lm_to_previous}}{A vector of labels from each datapoints to its closest previous landmarks.}

\item{\code{lm_weights}}{A vector of all landmark's weights.}

\item{\code{previous_to_current}}{A vector of previous to current label.}

\item{\code{area_of_infleunce}}{A sparse matrix of area of influence.}

\item{\code{best_representatives}}{A vector of each datapoint's best representative.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(
  scalenum,
  num_scales,
  tmatrix,
  lm_to_original,
  lm_to_previous,
  lm_weights,
  previous_to_current,
  area_of_influence
)}}{Initialize sub-scale.}

\item{\code{toString()}}{This method returns the string representation of a sub-scale.}
}}

